tests/requirements.txt
pytest==7.4.0
streamlit>=1.26.0
pytesseract>=0.3.10
Pillow>=10.0.0
seleniumbase>=4.15.0
gpt4all>=2.0.0